refactor: move subroutes to own modules
